Move media components into their own component
